validations, second ask for deletes in all control panel for admin

+++ ESHIRT/client/src/Components/Admin/RecycleBin/RecycleBinDesigns.jsx
@@ -32,9 +32,15 @@ export default function RecycleBinDesigns(){
 
     useEffect(() => {
       dispatch(getShirts());
-    }, []);
+    }, [count]);
 
     function handleDelete(e) {
+      swal({
+        title: "ARE YOU SURE?",
+        text: "The desing is going to delete definitely",
+        buttons: ["CANCEL", "DELETE"]
+      }).then(respuesta =>{
+        if(respuesta){
         dispatch(deleteShirt(parseInt(e.target.value))); 
         setCount(count +1);
         dispatch(getShirts())
@@ -46,7 +52,7 @@ export default function RecycleBinDesigns(){
           padding: "0.75rem"
           });
           dispatch(getShirts())
-      };
+      };})}
 
       function handlePublic(e) {
         const value = e.target.value;
